Tests: Restore Autogen/SameName .hh header test
In commit 399f9750ec4b4be6758306bfb45f9296d6426bf4, the Autogen/SameName test
was changed to not test .hh header file processing in AUTOGEN anymore,
because we had to revert that feature.  Since we now support .hh header file
processing in AUTOMOC again, we can restore the Autogen/SameName test for
that.