CPack/NSIS: Avoid "setup loading" screen by fixing template order
NSIS packages files (including those required by the installer itslef)
in the order in which they're mentioned in the .nsi file.

If solid compression is used, the installer might need to extract
large parts of the archive to be able to even start the setup program
("Please wait while setup is loading...").

Avoid this by moving component installation towards the end of the
template file.