Tests: Enable Fortran submodule tests on XL compilers
Since commit b66bc6606e (Tests: Add Fortran submodule tests, 2016-09-22,
v3.7.0-rc1~55^2) we try a small test program to see if the Fortran
compiler supports submodules.  However, a typo in the test program
caused it to fail on XL with the error:

    1513-083 (E) Internal or module function id was not set within the function.

Fix the typo so that the check passes and enables the submodule tests
with XL compilers.