FindSquish: Remove settingsGroup argument
This argument never worked and was not passed on Windows, which is why
the Windows version needs fewer fixups (and used to work independent of
this change).
On Linux (and macOS) it was passed and prevented the server from starting and the
test would not be able to run.

See also comments on https://gitlab.kitware.com/cmake/cmake/-/merge_requests/4648