Apple: suppress Xcode 12 ranlib warning
The Xcode 12 (Beta) toolchain for arm64 emits a warning when static
libraries are created:

```
ranlib: archive member: libStaticLib.a(a.c.o) offset in archive not a multiple of 8 (must be since member is an 64-bit object file)
```

Due to the unexpected output multiple tests are broken. To make those
tests pass we filter the problematic output in the test driver.

See: https://developer.apple.com/forums/thread/652620